+++ example/lib/wave_data/wave_data.dart
@@ -7,6 +7,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_soloud/flutter_soloud.dart';
 import 'package:logging/logging.dart';
 
+/// Waveform example.
+///
+/// This example demonstrates how read audio samples from a wave file.
 void main() async {
   // The `flutter_soloud` package logs everything
   // (from severe warnings to fine debug messages)
@@ -68,7 +71,7 @@ class _HelloFlutterSoLoudState extends State<HelloFlutterSoLoud> {
           children: [
             const SizedBox(height: 16),
             const Text(
-              'Read N audio samples (here N = width) from audio file.',
+              'Read N audio samples (here N = width*4) from audio file.',
             ),
             OutlinedButton(
               onPressed: () async {