+++ lib/src/bindings/bindings_player.dart
@@ -4,6 +4,7 @@ import 'dart:typed_data';
 import 'package:flutter_soloud/src/bindings/audio_data.dart';
 import 'package:flutter_soloud/src/enums.dart';
 import 'package:flutter_soloud/src/filters/filters.dart';
+import 'package:flutter_soloud/src/helpers/playback_device.dart';
 import 'package:flutter_soloud/src/sound_handle.dart';
 import 'package:flutter_soloud/src/sound_hash.dart';
 import 'package:meta/meta.dart';
@@ -52,6 +53,7 @@ abstract class FlutterSoLoud {
 
   /// Initialize the player. Must be called before any other player functions.
   ///
+  /// [deviceId] the device ID. -1 for default OS output device.
   /// [sampleRate] the sample rate. Usually is 22050, 44100 (CD quality)
   /// or 48000.
   /// [bufferSize] the audio buffer size. Usually is 2048, but can be also be
@@ -61,10 +63,20 @@ abstract class FlutterSoLoud {
   /// Returns [PlayerErrors.noError] if success.
   @mustBeOverridden
   PlayerErrors initEngine(
+    int deviceId,
     int sampleRate,
     int bufferSize,
     Channels channels,
   );
+
+  /// Change the playback device.
+  ///
+  /// [deviceId] the device ID. -1 for default OS output device.
+  @mustBeOverridden
+  PlayerErrors changeDevice(int deviceId);
+
+  /// List available playback devices.
+  List<PlaybackDevice> listPlaybackDevices();
 
   /// Must be called when the player is no more needed or when closing the app.
   @mustBeOverridden