+++ lib/src/enums.dart
@@ -62,7 +62,10 @@ enum PlayerErrors {
   soundHandleNotFound(17),
 
   /// Error getting filter parameter.
-  filterParameterGetError(18);
+  filterParameterGetError(18),
+
+  /// No playback devices were found.
+  noPlaybackDevicesFound(19);
 
   const PlayerErrors(this.value);
 
@@ -117,6 +120,9 @@ enum PlayerErrors {
       case PlayerErrors.filterParameterGetError:
         return 'An error (nan or inf value) occurred while getting a '
             'filter parameter!';
+      case PlayerErrors.noPlaybackDevicesFound:
+        return 'No playback devices were found while initializing engine or '
+            'when changing the output device.';
     }
   }